move resize & rotate helpers to UI class
