+++ src/tests/helpers/ui.ts
@@ -9,6 +9,10 @@ import type {
   ExcalidrawDiamondElement,
   ExcalidrawTextContainer,
 } from "../../element/types";
+import {
+  getTransformHandles,
+  TransformHandleDirection,
+} from "../../element/transformHandles";
 import { KEYS } from "../../keys";
 import { type ToolName } from "../queries/toolQueries";
 import { fireEvent, GlobalTestState, screen } from "../test-utils";
@@ -405,6 +409,46 @@ export class UI {
     fireEvent.input(editor, { target: { value: newText } });
     await new Promise((resolve) => setTimeout(resolve, 0));
     editor.blur();
+  }
+
+  static resize(
+    element: ExcalidrawElement,
+    handleDir: TransformHandleDirection,
+    mouseMove: [number, number],
+    keyboardModifiers: KeyboardModifiers = {},
+  ) {
+    mouse.select(element);
+    const handle = getTransformHandles(element, h.state.zoom, "mouse")[
+      handleDir
+    ]!;
+    const clientX = handle[0] + handle[2] / 2;
+    const clientY = handle[1] + handle[3] / 2;
+    Keyboard.withModifierKeys(keyboardModifiers, () => {
+      mouse.reset();
+      mouse.down(clientX, clientY);
+      mouse.move(mouseMove[0], mouseMove[1]);
+      mouse.up();
+    });
+  }
+
+  static rotate(
+    element: ExcalidrawElement,
+    deltaX: number,
+    deltaY: number,
+    keyboardModifiers: KeyboardModifiers = {},
+  ) {
+    mouse.select(element);
+    const handle = getTransformHandles(element, h.state.zoom, "mouse")
+      .rotation!;
+    const clientX = handle[0] + handle[2] / 2;
+    const clientY = handle[1] + handle[3] / 2;
+
+    Keyboard.withModifierKeys(keyboardModifiers, () => {
+      mouse.reset();
+      mouse.down(clientX, clientY);
+      mouse.move(clientX + deltaX, clientY + deltaY);
+      mouse.up();
+    });
   }
 
   static group(elements: ExcalidrawElement[]) {
